Перенесет функции для рендеринга из `utils.js` в `utils/render.js`
